+++ packages/web/src/app/[lng]/(main)/community/[articleId]/components/CommentForm.tsx
@@ -14,7 +14,7 @@ import { useTranslation } from '@/app/i18n/client';
 import { Icon } from '@/components/Icon';
 import { TextFieldController } from '@/components/TextField';
 import { useNumberParams } from '@/hooks/useNumberParams';
-import cn from '@/utils/cn';
+import { cn } from '@/utils';
 
 export type CreateCommentRequest = {
   content: string;
