+++ packages/web/src/app/[lng]/(main)/community/components/ArticleItem.client.tsx
@@ -13,7 +13,7 @@ import { Icon } from '@/components/Icon';
 import { Flex } from '@/components/Layout';
 import { Spacing } from '@/components/Spacing';
 import useAppRouter from '@/hooks/useAppRouter';
-import cn from '@/utils/cn';
+import { cn } from '@/utils';
 import { currentKoreaTime } from '@/utils/date';
 
 const formatDate = (date: string, locale: Locale) => {
