+++ packages/web/src/components/Divider/Divider.tsx
@@ -1,6 +1,6 @@
 import { memo } from 'react';
 
-import cn from '@/utils/cn';
+import { cn } from '@/utils';
 
 interface DividerProps<T extends React.ElementType> {
   /**
