+++ packages/web/src/components/Footer/Footer.tsx
@@ -1,5 +1,3 @@
-import { useTranslation } from 'react-i18next';
-
 import { ButtonAnimation } from '../Animation';
 import { Icon } from '../Icon';
 import { NavLink } from '../NavLink';
@@ -7,7 +5,7 @@ import { NavLink } from '../NavLink';
 import type { PageType } from '@/types';
 
 import { serverTranslation } from '@/app/i18n';
-import cn from '@/utils/cn';
+import { cn } from '@/utils';
 
 interface TabType {
   id: string;
