+++ packages/web/src/components/Modal/ModalWrapper.client.tsx
@@ -6,7 +6,7 @@ import type { StrictPropsWithChildren } from '@/types';
 
 import { fadeInVariants } from '@/constants/motions';
 import { useOnClickOutside } from '@/hooks/useOnClickOutside';
-import cn from '@/utils/cn';
+import { cn } from '@/utils';
 
 interface ModalWrapperProps {
   onClose?: () => void;
